% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.phi.R
\name{power.phi}
\alias{power.phi}
\title{Statistical power of phi-divergence test.}
\usage{
power.phi(alpha, n, s, beta, method = "gaussian-gaussian", eps = 0,
  mu = 0, df = 1, delta = 0)
}
\arguments{
\item{alpha}{- type-I error rate.}

\item{n}{- dimension parameter, i.e. the number of input statitics to construct phi-divergence statistic.}

\item{s}{- phi-divergence parameter. s = 2 is the higher criticism statitic.s = 1 is the Berk and Jones statistic.}

\item{beta}{- search range parameter . Beta must be between 1/n and 1.}

\item{method}{- different alternative hypothesis, including mixtures such as, "gaussian-gaussian", "gaussian-t", "t-t", "chisq-chisq", and "exp-chisq". By default, we use Gaussian mixture.}

\item{eps}{- mixing parameter of the mixture.}

\item{mu}{- mean of non standard Gaussian model.}

\item{df}{- degree of freedom of t/Chisq distribution and exp distribution.}

\item{delta}{- non-cenrality of t/Chisq distribution.}
}
\value{
Power of phi-divergence test.
}
\description{
Statistical power of phi-divergence test.
}
\details{
We consider the following hypothesis test,
\deqn{H_0: X_i\sim F, H_a: X_i\sim G}
Specifically, \eqn{F = F_0} and \eqn{G = (1-\epsilon)F_0+\epsilon F_1}, where \eqn{\epsilon} is the mixing parameter, \eqn{F_0} and \eqn{F_1} is
speified by the "method" argument:

"gaussian-gaussian": \eqn{F_0} is the standard normal CDF and \eqn{F = F_1} is the CDF of normal distribution with \eqn{\mu} defined by mu and \eqn{\sigma = 1}.

"gaussian-t": \eqn{F_0} is the standard normal CDF and \eqn{F = F_1} is the CDF of t distribution with degree of freedom defined by df.

"t-t": \eqn{F_0} is the CDF of t distribution with degree of freedom defined by df and \eqn{F = F_1} is the CDF of non-central t distribution with degree of freedom defined by df and non-centrality defined by delta.

"chisq-chisq": \eqn{F_0} is the CDF of Chisquare distribution with degree of freedom defined by df and \eqn{F = F_1} is the CDF of non-central Chisquare distribution with degree of freedom defined by df and non-centrality defined by delta.

"exp-chisq": \eqn{F_0} is the CDF of exponential distribution with parameter defined by df and \eqn{F = F_1} is the CDF of non-central Chisqaure distribution with degree of freedom defined by df and non-centrality defined by delta.
}
\examples{
pval <- runif(10)
hcstat <- stat.phi(pval, 2, 0.5)$value
alpha_phi <- 1 - pphi(q=hcstat, n=10, s=2, beta=0.5)
#If the alternative hypothesis Gaussian mixture with eps = 0.1 and mu = 1.2:#
power.phi(alpha_phi, 10, 2, 0.5, eps = 0.1, mu = 1.2)
}
\references{
1. Hong Zhang, Jiashun Jin and Zheyang Wu. "Distributions and Statistical Power of Optimal
Signal Detection Methods in Finite Samples", submitted.

2. Donoho, David; Jin, Jiashun. "Higher criticism for detecting sparse heterogeneous mixtures". Annals of Statistics 32 (2004).
}
\seealso{
\code{\link{stat.phi}} for the definition of the statistic.
}

